+++ examples/java/cdap/README.md
@@ -1,0 +1,24 @@
+<!--
+    Licensed to the Apache Software Foundation (ASF) under one
+    or more contributor license agreements.  See the NOTICE file
+    distributed with this work for additional information
+    regarding copyright ownership.  The ASF licenses this file
+    to you under the Apache License, Version 2.0 (the
+    "License"); you may not use this file except in compliance
+    with the License.  You may obtain a copy of the License at
+      http://www.apache.org/licenses/LICENSE-2.0
+    Unless required by applicable law or agreed to in writing,
+    software distributed under the License is distributed on an
+    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+    KIND, either express or implied.  See the License for the
+    specific language governing permissions and limitations
+    under the License.
+-->
+
+# Apache Beam pipeline examples for CdapIO and CDAP plugins
+
+This directory contains set of [Apache Beam](https://beam.apache.org/) pipeline examples to read data
+from a [CDAP plugin](https://github.com/data-integrations) and write data into .txt file (and vice versa).
+
+Supported CDAP plugins:
+- [ServiceNow](https://github.com/data-integrations/servicenow-plugins)
