+++ examples/java/cdap/src/main/java/org/apache/beam/examples/complete/cdap/utils/StructuredRecordUtils.java
